Add chat log colors, timestamp, filter & emoji box
Chat log names are now based off the ponies coat and mane.
Added a timestamp infront of each message.

There now is a filter on the top right of the chat box, with it you
can filter the past log.

Use no prefix for normal non-case sensitive search,
use # for case sensitive search and finally
use / for RegExp search.

When party or whisper messages are received, the respective tab is
highlighted.

Added an emoji picker, it is displayed when the chat bar is activated.

Co-authored-by: Stubenhocker1399 <Stubenhocker1399@gmail.com>